refactor(pv-scripts): extracted getCustomWebpackConfig for path helpers to webpackConfigHelpers
